+++ src/main/kotlin/com/awkrid/todo/domain/todo/service/TodoServiceImpl.kt
@@ -7,11 +7,14 @@ import com.awkrid.todo.domain.todo.dto.UpdateTodoRequest
 import com.awkrid.todo.domain.todo.model.Todo
 import com.awkrid.todo.domain.todo.model.toResponse
 import com.awkrid.todo.domain.todo.repository.TodoRepository
+import com.awkrid.todo.domain.user.exception.InvalidCredentialException
 import com.awkrid.todo.domain.user.repository.UserRepository
+import com.awkrid.todo.infra.swagger.security.UserPrincipal
 import jakarta.transaction.Transactional
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.core.Authentication
 import org.springframework.stereotype.Service
 
 @Service
@@ -31,8 +34,10 @@ class TodoServiceImpl(
     }
 
     @Transactional
-    override fun createTodo(request: CreateTodoRequest): TodoResponse {
-        val user = userRepository.findByIdOrNull(request.userId) ?: throw ModelNotFoundException("User", request.userId)
+    override fun createTodo(request: CreateTodoRequest, authentication: Authentication): TodoResponse {
+        val userPrincipal = authentication.principal as UserPrincipal
+        val user =
+            userRepository.findByIdOrNull(userPrincipal.id) ?: throw ModelNotFoundException("User", userPrincipal.id)
         val todo = Todo(
             title = request.title,
             description = request.description,
@@ -42,15 +47,23 @@ class TodoServiceImpl(
     }
 
     @Transactional
-    override fun updateTodo(todoId: Long, request: UpdateTodoRequest): TodoResponse {
+    override fun updateTodo(todoId: Long, request: UpdateTodoRequest, authentication: Authentication): TodoResponse {
+        val userPrincipal = authentication.principal as UserPrincipal
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
+        if (todo.user.id != userPrincipal.id) {
+            throw InvalidCredentialException()
+        }
         todo.updateTodo(request)
         return todoRepository.save(todo).toResponse()
     }
 
     @Transactional
-    override fun deleteTodo(todoId: Long) {
+    override fun deleteTodo(todoId: Long, authentication: Authentication) {
+        val userPrincipal = authentication.principal as UserPrincipal
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
+        if (todo.user.id != userPrincipal.id) {
+            throw InvalidCredentialException()
+        }
         todoRepository.delete(todo)
     }
